+++ src/main/java/com/kakaotech/team14backend/point/application/PointService.java
@@ -1,14 +1,12 @@
-package com.kakaotech.team14backend.outer.point.service;
+package com.kakaotech.team14backend.point.application;
 
-import com.kakaotech.team14backend.post.exception.PostNotFoundException;
 import com.kakaotech.team14backend.member.domain.Member;
-import com.kakaotech.team14backend.inner.point.model.UsePointDecider;
-import com.kakaotech.team14backend.inner.point.usecase.UsePointUsecase;
-import com.kakaotech.team14backend.inner.point.usecase.ValidatePointByPopularPostUsecase;
-import com.kakaotech.team14backend.post.domain.Post;
-import com.kakaotech.team14backend.post.infrastructure.PostRepository;
+import com.kakaotech.team14backend.point.domain.UsePointDecider;
+import com.kakaotech.team14backend.point.dto.UsePointByPopularPostRequestDTO;
 import com.kakaotech.team14backend.post.application.SetPostInstaCountUsecase;
-import com.kakaotech.team14backend.outer.point.dto.UsePointByPopularPostRequestDTO;
+import com.kakaotech.team14backend.post.domain.Post;
+import com.kakaotech.team14backend.post.exception.PostNotFoundException;
+import com.kakaotech.team14backend.post.infrastructure.PostRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
